requirements.txt
aiogram==3.23.0
aiohttp==3.13.0
aiosqlite==0.21.0
openai==1.109.1
pydantic==2.10.6
python-dotenv==1.0.0
